Fix integer overflow in mp_check and mp_next for MP_MAP
An MP_MAP header encodes the number of pairs stored in the map so we
multiply the value by 2 in mp_next and mp_check. The problem is that we
don't extend the integer value before peforming multiplication so it can
lead to an integer overflow and, as a result, mp_check fails to detect
invalid MsgPack data. Fix this.

Note, we only add a test for mp_check, because adding a test for mp_next
would make it consume too much memory (> 4GB).

Needed for https://github.com/tarantool/security/issues/18

+++ msgpuck.h
@@ -2483,11 +2483,11 @@ mp_next_slowpath(const char **data, int64_t k)
 			break;
 		case MP_HINT_MAP_16:
 			/* MP_MAP (16) */
-			k += 2 * mp_load_u16(data);
+			k += 2 * (uint32_t)mp_load_u16(data);
 			break;
 		case MP_HINT_MAP_32:
 			/* MP_MAP (32) */
-			k += 2 * mp_load_u32(data);
+			k += 2 * (uint64_t)mp_load_u32(data);
 			break;
 		case MP_HINT_EXT_8:
 			/* MP_EXT (8) */
@@ -2610,12 +2610,12 @@ mp_check(const char **data, const char *end)
 		case MP_HINT_MAP_16:
 			/* MP_MAP (16) */
 			MP_CHECK_LEN(sizeof(uint16_t));
-			k += 2 * mp_load_u16(data);
+			k += 2 * (uint32_t)mp_load_u16(data);
 			break;
 		case MP_HINT_MAP_32:
 			/* MP_MAP (32) */
 			MP_CHECK_LEN(sizeof(uint32_t));
-			k += 2 * mp_load_u32(data);
+			k += 2 * (uint64_t)mp_load_u32(data);
 			break;
 		case MP_HINT_EXT_8:
 			/* MP_EXT (8) */
